requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.18
certifi==2018.4.16
chardet==3.0.4
Django==2.1
django-haystack==2.8.1
idna==2.7
jieba==0.39
kombu==3.0.37
lxml==4.2.4
pytz==2018.5
redis==2.10.3
requests==2.19.1
urllib3==1.23
Whoosh==2.7.4
